Imported PaymentAggregator and moved shared code between APIGateway and Payment Aggregator to mApi Common.

+++ src/MerchantAPI/APIGateway/APIGateway.Rest/ViewModels/FeeAmountViewModelGet.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2020 Bitcoin Association
 
-using MerchantAPI.APIGateway.Domain.Models;
+using MerchantAPI.Common.Domain.Models;
 using System.Text.Json.Serialization;
 
 namespace MerchantAPI.APIGateway.Rest.ViewModels
@@ -8,10 +8,10 @@ namespace MerchantAPI.APIGateway.Rest.ViewModels
   public class FeeAmountViewModelGet
   {
     [JsonPropertyName("satoshis")]
-    public long Satoshis { get; set; }
+    public int Satoshis { get; set; }
 
     [JsonPropertyName("bytes")]
-    public long Bytes { get; set; }
+    public int Bytes { get; set; }
 
     public FeeAmountViewModelGet() { }
